// src/shift/dto/UpdateShift.dto.ts
export class UpdateShiftDto {
    shift_code: string;
    date?: string;
    shift_time?: string;
    product_target?: string;
    success_product_in_shift_time?: string;
    success_product_in_OT_time?: string;
    ideal_performance?: string;
    all_member?: number;
    checkin_member?: number;
    department_id?: string;
}